Allow v1 addons to rely on renaming system even when they have an explicit dep
An explicit dependency in package.json is supposed to take precedence over the package renaming system. But v1 addons without ember-auto-import don't actually get to access their explicit package.json dependencies from runtime code, so in that particular situation the renaming rules should still apply.

This was essentialy an asymetry between the check we do here in handleRenaming and the later "v1 package without auto-import" check in prehandleExternal. If an import is going to hit that `external`, it should also have the chance to first hit the renaming rules, avoiding the external entirely in many cases.