fix release-plan unlabelled changes PR
release-plan is supposed to report unlabelled PRs in the Prepare Release pr but it hasn't been working correctly. I figured out that the issue is that the pull request action won't open a PR unless there are some file changes and `stderr.log` is covered by our .gitignore file.

This PR fixes that and it also makes sure that the stderr.txt doesn't make it into the PR when the release-plan job is successful.