Update test matrix
This switching our tests from "lts_5_8" to "release" and locks the release into 5.11 plus my bugfix backport PR https://github.com/emberjs/ember.js/pull/20743.

I'm doing this because the circularity problems continue to be obnoxious, and some of the things I need to test are 5.x behaviors that are removed at 6.x.