Mandatory staticEmberSource and no more AMD externals
The older AMD externals system isn't really interoperable with true ES modules (and thus vite). We've kept it around this far, but the only way to actually use it is to manually register every single export that needs to be found in AMD, which is pretty clunkky.

It's also not really needed. In a typical app, zero imports (actual import statements. The ember resolver is a separate topic.) are actually getting handled by the AMD loader.

This PR drops the staticEmberSource option by forcing it to always be true, and drops the amdCompatibility option because it's no longer allowed.