Adding forced v2 packages feature
This adds a feature to `@embroider/shared-internals` so that the env var `EMBROIDER_FORCED_V2_PACKAGES` can cause any package to be interpreted as a v2 package regardless of what the package itself says.

The purpose of this is to unblock testing of ember-source as a v2 addon. Current ember-source canary is formatted as a valid v2 addon except it doesn't declare itself to be one, because that would be potentially-breaking for people who are still consuming it via AMD and/or relying on the specific timing details of the classic vendor.js.