moduleResolution nodenext
This updates the monorepo tsconfig to use module: nodenext, moduleResolution: nodenext.

Under this setting, typescript respect the Node's ideas about which files are CJS vs ESM. Most of our modules still compile to CJS, this PR isn't doing that conversion.

This PR does convert the `@embroider/vite` package to build as ESM, because vite itself makes it impossible to consume it's types from a CJS packge, and `@embroider/vite`'s outbound deps aren't too bad. This does mean that when `@embroider/vite` imports from `@embroider/core` it needs to actually treat core as CJS, which is weird-looking but acceptable for now.