Fix broken default export when required from CJS
When this package was converted to TS, `module.exports = ` was converted to `export default`, without taking into account that this would TS emit this as a named export of "default" (`exports.default = sideWatch`), so CJS consumers would have to call it as `require('@embroider/broccoli-side-watch').default()` (aka "double default" problem). As we are only targeting CJS consumers, this changes makes the compiled js have a single main export as `module.exports = sideWatch`.