Fix hbs plugin not resolving .hbs due to broken Regex
The Regex in `syntheticJStoHBS` was broken.

Given a TO .hbs such as `components/json/index.hbs`, the hbs rollup would need to resolve that from a module id of `components/json/index.js` to the actual `components/json/index.hbs` file on disk. But with the broken regex, it tried instead to find `components.hbson/index.hbs`!

Added some unit tests (that failed without the fix) to prevent regressions.