Add a new option for addon-shim to pass config to ember-auto-import
This is part of the plan to ship ember-source as a v2 addon.

This piece makes it possible for a v2 addon to send additional configuration directly to the leading copy of ember-auto-import. This is different from how it worked before, which was that registerV2Addon calls all cascaded up the tree through intermediate copies of ember-auto-import, which might be any compatible minor version.

Instead, now if you add a specific version of ember-auto-import as a dependency of your v2 addon, you will be guaranteed to be talking to that version (or newer) when you registerV2Addon, and thus you can send new options and expect them to work.