Eliminate 'ignored' case from module resolver
The ignored result was created to deal with the weird way vite resolves many things as "external" during dependency prescanning. But it was still complicating our ability to properly detect components during prescan.

This refactor manages to use the state capture we were already relying on to identify true "not_found" that vite was ignoring to also provide the true "found" that vite is ignoring, so our own use of the resolver never needs to deal with an "ignored" state.