Fix vendor ember-template-compiler.js on v2 ember-source
This fixes a typo pointed out by @chancancode.

This code exists to allow people to `app.import('vendor/ember-template-compiler.js')` after upgrading to ember-source >= 6.2.

It's not clear whether anybody actually *does* that, and the forward-looking plan for this area is that v2 apps can't do `app.import`, rendering the question moot.