fix transform-dependent compat resolver tests
These tests were not testing the right thing. They're supposed to simulate the existence of an ast plugin that inserts `@` into component names (which is a real thing in the wild), but that special transform was running *after* our resolver transform and altering its results. We didn't happen to notice until an unrelated change in babel-import-util made those alterations breaking.

I also added a separate test to cover the case that this was accidentally covering: that we can emit JS identifiers containing `$` (which we can, and now will when using the latest babel-import-util).