Use stable vite by default in development
`@embroider/vite` gets `vite` via peer dependency, but in development we need to pick some version to put into devDeps. The rolldown betas break a lot here, so I'd rather we only test those in the one place in the matrix where we're deliverately doing that. It looks like the newer betas have better typescript types that no longer conflict with the stable vite types.