Fix dependency-aware macros in dummy apps
This fixes the issue in  https://github.com/embroider-build/embroider/pull/2628.

appEmberSatisfies and dependencySatisfies need to work inside classic dummy apps, and didn't.

There was already a passing skipped test for a classic build of an addon's dummy app that got skipped too agressively when we dropped support for dummy apps. This unskips it and extends it.